Created MagmaHeart.Core.Input and MagmaHeart.Core.CameraControls namespaces. Moved files to the appropriate directories
