+++ Assets/_Scripts/Core/CameraControls.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 97d42339d8391054bbfe17befe983a4b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
