+++ Assets/_Scripts/Core/CameraControls/CameraController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 5a8c924c5f7887145800681ce4005052
