+++ Assets/_Scripts/Core/CameraControls/CameraTargetTracker.cs
@@ -1,0 +1,4 @@
+namespace MagmaHeart.Core.CameraControls
+{
+
+}
