+++ Assets/_Scripts/Core/CameraControls/TurnBasedCameraBehaviour.cs
@@ -1,6 +1,7 @@
+using MagmaHeart.Core.Input;
 using UnityEngine;
 
-namespace MagmaHeart.Core.Entities.PlayableCharacters
+namespace MagmaHeart.Core.CameraControls
 {
     public class TurnBasedCameraBehaviour : ICameraBehaviour
     {
