+++ Assets/_Scripts/Core/CombatSystem/CombatStateSwitcher.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using MagmaHeart.Core.CameraControls;
 using MagmaHeart.Core.Dungeon;
 using MagmaHeart.Core.Entities.PlayableCharacters;
 using MagmaHeart.Core.UI;
