--- Assets/_Scripts/Core/Entities/PlayableCharacters/ActionUserInput.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: cbd2802b3b616d5428093d8ebb975314
