--- Assets/_Scripts/Core/Entities/PlayableCharacters/CameraController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4174dad3544a3db409c2da75aa073e2a
