--- Assets/_Scripts/Core/Entities/PlayableCharacters/MouseControl.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 89bf3325fbaf30244a74d0f250ea2997
