--- Assets/_Scripts/Core/Entities/PlayableCharacters/TurnBasedCameraBehaviour.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 94f2708d65a883b4895c83ced524f923
