--- Assets/_Scripts/Core/Entities/PlayableCharacters/TurnBasedUserInput.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5e549520e848b774b82c3dcb19cbf903
