+++ Assets/_Scripts/Core/Input.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4196419c65f3681479c7128c3bfea71a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
