+++ Assets/_Scripts/Core/Input/TurnBasedUserInput.cs
@@ -1,7 +1,7 @@
 using MagmaHeart.Core.Dungeon;
 using UnityEngine;
 
-namespace MagmaHeart.Core.Entities.PlayableCharacters
+namespace MagmaHeart.Core.Input
 {
     public class TurnBasedUserInput
     {
