--- Assets/_Scripts/Core/UserInput.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 3719a4714e2f6d746b8790b230147553
